+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is used to list changes made in each version of the gem.
 - Update redis configuration in generators.
 - Update mailcatcher docker image for better compatibility with ARM64.
 - Remove possible suffix `Service` in service generator.
+- Update rubocop configuration in template.
 
 # 1.2.1